refactor(Lotto): 파일 이동
